# Conflicts:
#	app/Http/Resources/UserResource.php
#	database/factories/CategoryFactory.php
#	lang/fa/general.php

+++ app/Actions/Notice/UpdateNoticeAction.php
@@ -2,7 +2,9 @@
 
 namespace App\Actions\Notice;
 
+use App\Actions\Translation\SetTranslationAction;
 use App\Models\Notice;
+use App\Repositories\Category\CategoryRepositoryInterface;
 use App\Repositories\Notice\NoticeRepositoryInterface;
 use Illuminate\Support\Facades\DB;
 use Lorisleiva\Actions\Concerns\AsAction;
@@ -11,7 +13,8 @@ class UpdateNoticeAction
 {
     use AsAction;
 
-    public function __construct(private readonly NoticeRepositoryInterface $repository)
+    public function __construct(private readonly NoticeRepositoryInterface $repository ,
+        private readonly CategoryRepositoryInterface $categoryRepository)
     {
     }
 
@@ -24,11 +27,23 @@ class UpdateNoticeAction
     public function handle(Notice $notice, array $payload): Notice
     {
         return DB::transaction(function () use ($notice, $payload) {
-            $notice->media()->delete();
-            $notice->addMediaFromRequest('media')
-                   ->toMediaCollection('notice');
-            $notice->update($payload);
-            return $notice;
+            $category = $this->categoryRepository->find($payload['category_id']);
+
+            if ($category->type == Notice::class) {
+                $payload['user_id'] = auth()->user()->id;
+                $model = $this->repository->update($notice, $payload);
+                $model->extra_attributes->set($payload['extra_attributes']);
+                $model->save();
+                SetTranslationAction::run($notice, $payload['translations']);
+
+                if (isset($payload['media'])) {
+                    $notice->media()->delete();
+                    $notice->addMediaFromRequest('media')
+                           ->toMediaCollection('notice');
+                }
+            }
+
+            return $notice->load('translations');
         });
     }
 }
